+++ datastore/datastore.go
@@ -1,0 +1,4 @@
+package datastore
+
+type DataStore interface {
+}
