--- drivers/driver.go
@@ -1,5 +1,0 @@
-package drivers
-
-type DatabaseDriver interface {
-	
-}
