+++ gext-server.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/owenoclee/gext-server/drivers"
+	"github.com/owenoclee/gext-server/datastore"
 )
 
 func main() {
 	db := initDB()
 	defer db.Close()
-	drivers.Initialize(db)
+	datastore.Initialize(db)
 	router := initRouter(db)
 	serveHTTP(router)
 }
